requirements/requirements.txt
absl-py>=0.13.0
attrs>=21.4.0
chex>=0.0.8
jax>=0.2.18
jaxlib>=0.1.69
numpy>=1.21.1
opt-einsum>=3.3.0
requests>=2.19.0
six>=1.16.0
torch>=2.5.0
toolz>=0.11.1